abstracted logic from quantity dialog
